Update job scheduling and user saving in several Python scripts
This commit includes changes to multiple scripts. We've updated the jobs.py to run the scheduled jobs more frequently. Furthermore, the initialization of the users array in birthday_bot.py has been changed to load existing users, and save the users in the array only when a new user is appended. Lastly, the 'when' field has been added to the config.py.example file.